Additional code cleanup and deletion
