+++ module/proto/peggy/v1/attestation.proto
@@ -10,7 +10,6 @@ option go_package = "github.com/althea-net/peggy/module/x/peggy/types";
 // ClaimType is the cosmos type of an event from the counterpart chain that can be handled
 enum ClaimType {
   option (gogoproto.goproto_enum_prefix)      = false;
-  option (gogoproto.goproto_enum_stringer)    = false;
 
   CLAIM_TYPE_UNKNOWN                          = 0 [(gogoproto.enumvalue_customname) = "CLAIM_TYPE_UNKNOWN"];
   CLAIM_TYPE_ETHEREUM_BRIDGE_DEPOSIT          = 1 [(gogoproto.enumvalue_customname) = "CLAIM_TYPE_ETHEREUM_BRIDGE_DEPOSIT"];
