+++ module/x/peggy/types/errors.go
@@ -5,7 +5,6 @@ import (
 )
 
 var (
-	// ErrInternal redacts the error message
 	ErrInternal                = sdkerrors.Register(ModuleName, 1, "internal")
 	ErrDuplicate               = sdkerrors.Register(ModuleName, 2, "duplicate")
 	ErrInvalid                 = sdkerrors.Register(ModuleName, 3, "invalid")
