+++ module/x/peggy/types/ethereum.go
@@ -11,19 +11,18 @@ import (
 )
 
 const (
-	PeggyDenomPrefix      = "peggy"
-	PeggyDenomSeperator   = "/"
-	PeggyDenomPrefixLen   = len(PeggyDenomPrefix)
-	PeggyDenomSepLen      = len(PeggyDenomSeperator)
+	// PeggyDenomPrefix indicates the prefix for all assests minted by this module
+	PeggyDenomPrefix = ModuleName
+
+	// PeggyDenomSeperator is the seperator for peggy denoms
+	PeggyDenomSeperator = "/"
+
+	// ETHContractAddressLen is the length of contract address strings
 	ETHContractAddressLen = 42
-	PeggyDenomLen         = PeggyDenomPrefixLen + PeggyDenomSepLen + ETHContractAddressLen
+
+	// PeggyDenomLen is the length of the denoms generated by the peggy module
+	PeggyDenomLen = len(PeggyDenomPrefix) + len(PeggyDenomSeperator) + ETHContractAddressLen
 )
-
-// EthereumAddressLength is the length of an ETH address: 20 bytes
-// const EthereumAddressLength = gethCommon.AddressLength
-
-// EthereumAddress defines a standard ethereum address
-// type EthereumAddress gethCommon.Address
 
 // EthAddrLessThan migrates the Ethereum address less than function
 func EthAddrLessThan(e, o string) bool {
@@ -37,6 +36,9 @@ func ValidateEthAddress(a string) error {
 	}
 	if !regexp.MustCompile("^0x[0-9a-fA-F]{40}$").MatchString(a) {
 		return fmt.Errorf("address(%s) doesn't pass regex", a)
+	}
+	if len(a) != ETHContractAddressLen {
+		return fmt.Errorf("address(%s) of the wrong length exp(%d) actual(%d)", a, len(a), ETHContractAddressLen)
 	}
 	return nil
 }
@@ -64,11 +66,6 @@ func (e *ERC20Token) ValidateBasic() error {
 	return nil
 }
 
-// AsVoucherCoin converts the data into a cosmos coin with peggy voucher denom.
-func (e *ERC20Token) AsVoucherCoin() sdk.Coin {
-	return e.PeggyCoin()
-}
-
 // Add adds one ERC20 to another
 // TODO: make this return errors instead
 func (e *ERC20Token) Add(o *ERC20Token) *ERC20Token {
@@ -84,26 +81,26 @@ func (e *ERC20Token) Add(o *ERC20Token) *ERC20Token {
 
 // ERC20FromPeggyCoin returns the ERC20 representation of a given peggy coin
 func ERC20FromPeggyCoin(v sdk.Coin) (*ERC20Token, error) {
-	if !IsPeggyCoin(v) {
-		return nil, fmt.Errorf("%s isn't a valid peggy coin", v.String())
+	if err := ValidatePeggyCoin(v); err != nil {
+		return nil, fmt.Errorf("%s isn't a valid peggy coin: %s", v.String(), err)
 	}
-	return &ERC20Token{Contract: strings.Split(v.Denom, "/")[1], Amount: v.Amount}, nil
+	return &ERC20Token{Contract: strings.Split(v.Denom, PeggyDenomSeperator)[1], Amount: v.Amount}, nil
 }
 
-// IsPeggyCoin returns true if a coin is a peggy representation of an ERC20 token
-func IsPeggyCoin(v sdk.Coin) bool {
-	spl := strings.Split(v.Denom, "/")
+// ValidatePeggyCoin returns true if a coin is a peggy representation of an ERC20 token
+func ValidatePeggyCoin(v sdk.Coin) error {
+	spl := strings.Split(v.Denom, PeggyDenomSeperator)
 	err := ValidateEthAddress(spl[1])
 	switch {
 	case len(spl) != 2:
-		return false
-	case spl[0] != "peggy":
-		return false
+		return fmt.Errorf("denom(%s) not valid, more seperators(%s) than expected", v.Denom, PeggyDenomSeperator)
+	case spl[0] != PeggyDenomPrefix:
+		return fmt.Errorf("denom prefix(%s) not equal to expected(%s)", spl[0], PeggyDenomPrefix)
 	case err != nil:
-		return false
+		return fmt.Errorf("error(%s) validating ethereum contract address", err)
 	case len(v.Denom) != PeggyDenomLen:
-		return false
+		return fmt.Errorf("len(denom)(%d) not equal to PeggyDenomLen(%d)", len(v.Denom), PeggyDenomLen)
 	default:
-		return true
+		return nil
 	}
 }
