+++ module/x/peggy/types/ethereum_signer.go
@@ -2,14 +2,14 @@ package types
 
 import (
 	"crypto/ecdsa"
-	"encoding/json"
-	"fmt"
 
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/ethereum/go-ethereum/crypto"
 )
 
-const signaturePrefix = "\x19Ethereum Signed Message:\n32"
+const (
+	signaturePrefix = "\x19Ethereum Signed Message:\n32"
+)
 
 // NewEthereumSignature creates a new signuature over a given byte array
 func NewEthereumSignature(hash []byte, privateKey *ecdsa.PrivateKey) ([]byte, error) {
@@ -57,74 +57,5 @@ func ValidateEthereumSignature(hash []byte, signature []byte, ethAddress string)
 		return sdkerrors.Wrap(ErrInvalid, "signature not matching")
 	}
 
-	return nil
-}
-
-var signTypeToNames = map[SignType]string{
-	SIGN_TYPE_ORCHESTRATOR_SIGNED_MULTI_SIG_UPDATE: "orchestrator_signed_multisig_update",
-	SIGN_TYPE_ORCHESTRATOR_SIGNED_WITHDRAW_BATCH:   "orchestrator_signed_withdraw_batch",
-}
-
-// AllSignTypes types that are signed with by the bridge multisig set
-var AllSignTypes = []SignType{SIGN_TYPE_ORCHESTRATOR_SIGNED_MULTI_SIG_UPDATE, SIGN_TYPE_ORCHESTRATOR_SIGNED_WITHDRAW_BATCH}
-
-// IsSignType returns if given sign type is supported
-func IsSignType(s SignType) bool {
-	for _, v := range AllSignTypes {
-		if s == v {
-			return true
-		}
-	}
-	return false
-}
-
-// SignTypeFromName returns the sign type given its string representation
-func SignTypeFromName(s string) (SignType, bool) {
-	for _, v := range AllSignTypes {
-		name, ok := signTypeToNames[v]
-		if ok && name == s {
-			return v, true
-		}
-	}
-	return SIGN_TYPE_UNKNOWN, false
-}
-
-// ToSignTypeNames given sign types, it returns their associated strings
-func ToSignTypeNames(s ...SignType) []string {
-	r := make([]string, len(s))
-	for i := range s {
-		r[i] = s[i].String()
-	}
-	return r
-}
-
-func (t SignType) String() string {
-	return signTypeToNames[t]
-}
-
-// Bytes implements bytes
-func (t SignType) Bytes() []byte {
-	return []byte{byte(t)}
-}
-
-// MarshalJSON implements proto.Message
-func (t SignType) MarshalJSON() ([]byte, error) {
-	return []byte(fmt.Sprintf("%q", t.String())), nil
-}
-
-// UnmarshalJSON implements proto.Message
-func (t *SignType) UnmarshalJSON(input []byte) error {
-	if string(input) == `""` {
-		return nil
-	}
-	var s string
-	if err := json.Unmarshal(input, &s); err != nil {
-		return err
-	}
-	c, exists := SignTypeFromName(s)
-	if !exists {
-		return sdkerrors.Wrap(ErrUnknown, "claim type")
-	}
-	*t = c
 	return nil
 }
