+++ module/x/peggy/types/genesis_test.go
@@ -13,15 +13,21 @@ func TestGenesisStateValidate(t *testing.T) {
 	}{
 		"default params": {src: DefaultGenesisState(), expErr: false},
 		"empty params":   {src: &GenesisState{Params: &Params{}}, expErr: false},
-		//"invalid params": {src: GenesisState{
-		//	Params: Params{},  // can't test currently
-		//}, expErr: true},
+		"invalid params": {src: &GenesisState{
+			Params: &Params{
+				PeggyId:            []byte("foo"),
+				ContractSourceHash: "laksdjflasdkfja",
+				EthereumAddress:    "invalid-eth-address",
+				BridgeChainId:      3279089,
+			},
+		}, expErr: true},
 	}
 	for msg, spec := range specs {
 		t.Run(msg, func(t *testing.T) {
 			err := spec.src.ValidateBasic()
 			if spec.expErr {
 				require.Error(t, err)
+				return
 			}
 			require.NoError(t, err)
 		})
