--- module/x/peggy/types/nonce.go
@@ -1,23 +1,0 @@
-package types
-
-import (
-	"encoding/binary"
-	"strconv"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-)
-
-// UInt64FromBytes create uint from binary big endian representation
-func UInt64FromBytes(s []byte) uint64 {
-	return binary.BigEndian.Uint64(s)
-}
-
-// UInt64Bytes uses the SDK byte marshaling to encode a uint64
-func UInt64Bytes(n uint64) []byte {
-	return sdk.Uint64ToBigEndian(n)
-}
-
-// UInt64FromString to parse out a uint64 for a nonce
-func UInt64FromString(s string) (uint64, error) {
-	return strconv.ParseUint(s, 10, 64)
-}
