--- module/x/peggy/types/pool.go
@@ -1,1 +1,0 @@
-package types
